# server/helpers/swagger.yaml
basePath: /
consumes:
    - application/json
    - multipart/form-data
...
info:
    description: |-
        the purpose of this application is to provide an application
        that is using plain go code to define an API

        This should demonstrate all the possible comment annotations
        that are available to turn go code into a fully compliant swagger 2.0 spec
    license:
        name: Apache-2.0
        url: http://www.apache.org/licenses/LICENSE-2.0.txt
    title: Meshery API.
    version: 0.4.27
paths:
...
    /api/filter/{id}:
        delete:
            description: 'Deletes a meshplay filter with ID: id'
            operationId: idDeleteMesheryFilter
            parameters:
                - description: id for a specific
                  format: uuid
                  in: path
                  name: id
                  required: true
                  type: string
                  x-go-name: ID
            responses:
                "200":
                    $ref: '#/responses/noContentWrapper'
            summary: Handle Delete for a Meshery Filter
            tags:
                - FiltersAPI
        get:
            description: Fetches the Meshery Filter with the given id
            operationId: idGetMesheryFilter
            parameters:
                - description: id for a specific
                  format: uuid
                  in: path
                  name: id
                  required: true
                  type: string
                  x-go-name: ID
            responses:
                "200":
                    $ref: '#/responses/meshplayFilterResponseWrapper'
            summary: Handle GET request for a Meshery Filter
            tags:
                - FiltersAPI
...
produces:
    - application/json
responses:
...
    meshplayFilterResponseWrapper:
        description: Returns a single meshplay filter
...
    noContentWrapper:
        description: ""
...
schemes:
    - http
security:
    - token:
        - '[]'
securityDefinitions:
    token:
        in: cookie
        name: token
        type: apiKey
swagger: "2.0"